refactored - extracted shared logic into ProfileHelper to reduce code duplication
